nw-bench: merged client and server into a single binary
