+++ benches/netbench/Cargo.toml
@@ -21,33 +21,17 @@ hermit = { path = "../../hermit", default-features = false }
 default = ["hermit/acpi", "hermit/pci", "hermit/smp", "hermit/tcp", "hermit/udp"]
 
 [[bin]]
-name = "tcp-server-bw"
-path = "src/rust-tcp-bw/server.rs"
+name = "tcp-bw"
+path = "src/rust-tcp-bw.rs"
 
 [[bin]]
-name = "tcp-client-bw"
-path = "src/rust-tcp-bw/client.rs"
+name = "tcp-latency"
+path = "src/rust-tcp-latency.rs"
 
 [[bin]]
-name = "tcp-server-latency"
-path = "src/rust-tcp-latency/server.rs"
+name = "udp-bw"
+path = "src/rust-udp-bw.rs"
 
 [[bin]]
-name = "tcp-client-latency"
-path = "src/rust-tcp-latency/client.rs"
-
-[[bin]]
-name = "udp-server-bw"
-path = "src/rust-udp-bw/server.rs"
-
-[[bin]]
-name = "udp-client-bw"
-path = "src/rust-udp-bw/client.rs"
-
-[[bin]]
-name = "udp-server-latency"
-path = "src/rust-udp-latency/server.rs"
-
-[[bin]]
-name = "udp-client-latency"
-path = "src/rust-udp-latency/client.rs"
+name = "udp-latency"
+path = "src/rust-udp-latency.rs"
