--- benches/netbench/src/rust-tcp-bw/client.rs
@@ -1,52 +1,0 @@
-use std::io::{self, Write};
-use std::net::TcpStream;
-
-use clap::Parser;
-#[cfg(target_os = "hermit")]
-use hermit as _;
-use rust_tcp_io_perf::config::Config;
-use rust_tcp_io_perf::connection;
-use rust_tcp_io_perf::print_utils::ProgressPrinter;
-
-fn send_rounds(stream: &mut TcpStream, rounds: usize, bytes: usize, progress_print: bool) {
-	// Create a buffer of 0s, size n_bytes, to be sent over multiple times
-	let buf = vec![0; bytes];
-
-	let progress_printer = ProgressPrinter::new(rounds, progress_print);
-
-	for i in 0..rounds {
-		progress_printer.print(i);
-		let mut pos = 0;
-
-		while pos < buf.len() {
-			let bytes_written = match stream.write(&buf[pos..]) {
-				Ok(len) => len,
-				Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => 0,
-				Err(e) => panic!("encountered IO error: {e}"),
-			};
-			pos += bytes_written;
-		}
-	}
-	stream.flush().expect("Unexpected behaviour");
-}
-
-fn main() {
-	let args = Config::parse();
-
-	println!("Connecting to the server {}:{}...", args.address, args.port);
-
-	if let Ok(mut stream) = connection::client_connect(args.address_and_port()) {
-		connection::setup(&args, &stream);
-		println!("Connection established! Ready to send...");
-
-		send_rounds(&mut stream, args.warmup, args.n_bytes, false);
-		send_rounds(&mut stream, args.n_rounds, args.n_bytes, true);
-
-		stream.flush().expect("Unexpected behaviour");
-		connection::close_connection(&stream);
-
-		println!("Sent everything!");
-	} else {
-		println!("Couldn't connect to server...");
-	}
-}
